fix immutability done wrong within reducer initial state
The ADD_WORKSPACE_SUCCESS test points out that the initial state does not have immutable values for objects. We thought that makeTypedFactory returned a Record but in fact it is just used to type it.

In order to fix it, we have to redefine the makeTypedFactory function, see https://github.com/rangle/typed-immutable-record/issues/23