refactor(frontend): rework workspace components hierarcy
Relocate workspace management in workspace component, overview in
workspace-overview, workspaces dialog triggering in its own component
and move workspace list into workspaces folder.

This allows to simply a lot of code that is now only trigerred when it
should be (e.g. selectors for the workspace or the sidebar only when on
a workspace).
Also now we can use the router to leave a workspace directly.

This is in preparation for angular 4.2 which is stricter with routing.